backend/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
geographiclib==1.52
geopy==2.2.0
idna==3.3
requests==2.27.1
six==1.16.0
spotipy==2.19.0
urllib3==1.26.9
Flask==1.0.2
flask-Cors==3.0.8
flask-marshmallow==0.9.0
google-auth==1.27.0
google-auth-oauthlib==0.4.2
oauthlib==3.1.0
pathlib==1.0.1
requests==2.25.1
requests-oauthlib==1.3.0
pyjwt==1.7.1